+++ src/PayPalService.php
@@ -4,7 +4,9 @@ namespace PaymentGateway\PayPalSdk;
 
 use EasyHttp\GuzzleLayer\GuzzleClient;
 use EasyHttp\LayerContracts\Contracts\HttpClientResponse;
+use PaymentGateway\PayPalSdk\Requests\StorePlanRequest;
 use PaymentGateway\PayPalSdk\Requests\StoreProductRequest;
+use PaymentGateway\PayPalSdk\Requests\UpdatePlanRequest;
 use PaymentGateway\PayPalSdk\Requests\UpdateProductRequest;
 
 class PayPalService
@@ -66,11 +68,53 @@ class PayPalService
         return $this->client->execute();
     }
 
-    public function updateProduct(UpdateProductRequest $product): HttpClientResponse
+    public function getProduct(string $id): HttpClientResponse
     {
-        $this->client->prepareRequest('PATCH', $this->baseUri . '/v1/catalogs/products/' . $product->getId());
+        $this->client->prepareRequest('GET', $this->baseUri . '/v1/catalogs/products/' . $id);
         $this->client->getRequest()->setHeader('Authorization', 'Bearer ' . $this->getToken()['access_token']);
-        $this->client->getRequest()->setJson($product->toArray());
+
+        return $this->client->execute();
+    }
+
+    public function updateProduct(UpdateProductRequest $productRequest): HttpClientResponse
+    {
+        $this->client->prepareRequest('PATCH', $this->baseUri . '/v1/catalogs/products/' . $productRequest->getId());
+        $this->client->getRequest()->setHeader('Authorization', 'Bearer ' . $this->getToken()['access_token']);
+        $this->client->getRequest()->setJson($productRequest->toArray());
+
+        return $this->client->execute();
+    }
+
+    public function createPlan(StorePlanRequest $storePlanRequest): HttpClientResponse
+    {
+        $this->client->prepareRequest('POST', $this->baseUri . '/v1/billing/plans');
+        $this->client->getRequest()->setHeader('Authorization', 'Bearer ' . $this->getToken()['access_token']);
+        $this->client->getRequest()->setJson($storePlanRequest->toArray());
+
+        return $this->client->execute();
+    }
+
+    public function getPlans(): HttpClientResponse
+    {
+        $this->client->prepareRequest('GET', $this->baseUri . '/v1/billing/plans');
+        $this->client->getRequest()->setHeader('Authorization', 'Bearer ' . $this->getToken()['access_token']);
+
+        return $this->client->execute();
+    }
+
+    public function getPlan(string $id): HttpClientResponse
+    {
+        $this->client->prepareRequest('GET', $this->baseUri . '/v1/billing/plans/' . $id);
+        $this->client->getRequest()->setHeader('Authorization', 'Bearer ' . $this->getToken()['access_token']);
+
+        return $this->client->execute();
+    }
+
+    public function updatePlan(UpdatePlanRequest $planRequest): HttpClientResponse
+    {
+        $this->client->prepareRequest('PATCH', $this->baseUri . '/v1/billing/plans/' . $planRequest->getId());
+        $this->client->getRequest()->setHeader('Authorization', 'Bearer ' . $this->getToken()['access_token']);
+        $this->client->getRequest()->setJson($planRequest->toArray());
 
         return $this->client->execute();
     }