+++ src/Subscriptions/Constants/CurrencyCode.php
@@ -1,0 +1,33 @@
+<?php
+
+namespace PaymentGateway\PayPalSdk\Subscriptions\Constants;
+
+interface CurrencyCode
+{
+    public const AUSTRALIAN_DOLLAR = 'AUD';
+    public const BRAZILIAN_REAL = 'BRL';
+    public const CANADIAN_DOLLAR = 'CAD';
+    public const CHINESE_RENMENBI = 'CNY';
+    public const CZECH_KORUNA = 'CZK';
+    public const DANISH_KRONE = 'DKK';
+    public const EURO = 'EUR';
+    public const HONG_KONG_DOLLAR = 'HKD';
+    public const HUNGARIAN_FORINT = 'HUF';
+    public const INDIAN_RUPEE = 'INR';
+    public const ISRAELI_NEW_SHEKEL = 'ILS';
+    public const JAPANESE_YEN = 'JPY';
+    public const MALAYSIAN_RINGGIT = 'MYR';
+    public const MEXICAN_PESO = 'MXN';
+    public const NEW_TAIWAN_DOLLAR = 'TWD';
+    public const NEW_ZEALAND_DOLLAR = 'NZD';
+    public const NORWEGIAN_KRONE = 'NOK';
+    public const PHILIPPINE_PESO = 'PHP';
+    public const POLISH_ZLOTY = 'PLN';
+    public const POUND_STERLING = 'GBP';
+    public const RUSSIAN_RUBLE = 'RUB';
+    public const SINGAPORE_DOLLAR = 'SGD';
+    public const SWEDISH_KRONA = 'SEK';
+    public const SWISS_FRANC = 'CHF';
+    public const THAI_BAHT = 'THB';
+    public const UNITED_STATES_DOLLAR = 'USD';
+}