+++ CHANGELOG.md
@@ -4,3 +4,9 @@
 
 ### Fixed
 - Fixed request overlapping ([#18](https://github.com/payment-gateways/paypal-sdk/pull/19))
+
+## [v0.2.0 (2020-12-20)](https://github.com/laravel/framework/compare/v0.1.1...v0.2.0)
+
+### Added
+- Added `getProduct()` method in Catalog Products API ([#20](https://github.com/payment-gateways/paypal-sdk/pull/20))
+- Added Subscriptions API integration ([#20](https://github.com/payment-gateways/paypal-sdk/pull/20))